fix all register controller bug

+++ src/utils/cloudinary.js
@@ -1,26 +1,32 @@
-import {v2 as cloudinary} from "cloudinary"
+import { v2 as cloudinary } from "cloudinary"
 import fs from "fs"
 
-cloudinary.config({ 
-    cloud_name: process.env.CLODINARY_CLOUD_NAME, 
-    api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET 
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null;
+        if (!localFilePath) return null;
         //upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type:"auto"
+            resource_type: "auto"
         })
         //file has been successfully uploaded
-        console.log("File is uploaded on cloudinary",response.url);
+        //console.log("File is uploaded on cloudinary", response.url);
+        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved file as the upload operation got failed
+        if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    } // remove the locally saved file as the upload operation got failed
         return null;
     }
 }
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
